[in progress: 1315] SSO/AD: ha nem találjuk a VSQAccess AD usert akkor is beengedjük, csoport torles letiltas ha permament

+++ modules/Visitor/Groups/Controller.php
@@ -30,8 +30,11 @@ class Controller extends \Visitor\Controller {
       'groups',
       $this->application->getNumericParameter('id')
     );
-    $groupModel->delete( $groupModel->id );
-    
+
+    // only delete if the group not permanent
+    if ( !$groupModel->row['ispermanent'] )
+      $groupModel->delete( $groupModel->id );
+
     $this->redirect(
       $this->application->getParameter('forward', 'groups/index' )
     );
